fix: code review notes
